cast frame time to float before division

+++ kiln/engine/Kiln.cc
@@ -54,7 +54,7 @@ void Kiln::run(KilnModule& module) {
 
   while(isRunning && module.hasSub()) {
     tickStartTime = SDL_GetTicks();
-    deltaTime = (tickStartTime - lastTickStartTime) / 1000;
+    deltaTime = (float)(tickStartTime - lastTickStartTime) / 1000.f;
     printf("DT: %fs, FT: %dms\n", deltaTime, tickStartTime - lastTickStartTime);
     
     while(this->coreManagement.inputManager.poll()) {
